Fix inline size bubbling with fixed width block. Improve intrinsic inline size calculation for text fragments.
These two fixes are related to the wikipedia metabug #2554.

They don't make the wikipedia page look better (they cause a slight regression in the top caption table), but they are prerequisites for fixing some of the other layout issues that remain.

Added reftests for each of the three cases I have come across that this patch solves.