Build fixes for Android windowing code
This begins porting the Android event loop to work with the inverted flow
control from #3761.  Unfortunately, GLUT does not give us enough control over
the event loop to really make this work, so this will build but it may not run
properly.  Our current plan is to get rid of GLUT and switch to Glutin in the
near future.

r? @pcwalton