Add a `./mach clean-snapshots` command to remove old Cargo and Rust.
Bootstrapping automatically downloads new Rust and Cargo snapshots
as needed into versioned directories,
but do not remove now-unused versions.
This is the desired behavior for `git bisect` to be usable.

However, this means that old version keep accumulating, taking up disk space.
This adds a mach command to remove snapshots other than the ones
currently being used.
It is never run automatically.

To be safe, the command defaults to only printing what would be removed,
and only removes stuff when run with a `-f` argument.