Travis CI building cleanup, enable caching
The only reason the Dockerfile was introduced is because the default
machines that Travis uses are based on Ubuntu 12.04, which has some very
old incompatible dependencies with Servo. Docker allowed use to use a
new version of Ubuntu, allowing us to compile with ease. I just learned
that they are currently beta testing 14.04 support:

http://docs.travis-ci.com/user/trusty-ci-environment/

This commit updates our Travis config to remove our dependency on Docker
and just build directly on the images, reducing some complexity and also
overhead of downloading Docker images.

In addition, this commit also enables caching of the .servo and .cargo
directories on Travis in an attempt to reduce build times.

http://docs.travis-ci.com/user/caching/#Arbitrary-directories