Remove pointless join_layout calls.
Code on the script thread can only observe self.layout_join_port being Some()
between the time it is set in force_reflow, and the join_layout call later in
that function, and no significant code is called in that code.

As join_layout does nothing when called if self.layout_join_port is None,
these two calls are useless.