compositing: Only make the compositor responsible for initiating reflow for CSS transitions and animations, not `requestAnimationFrame()` callbacks.
In the case of the latter, the script thread will kick off the reflow if
it's necessary, so there's no need for the compositor to do it.

Some pages, like nytimes.com, like to call `requestAnimationFrame()`
without actually mutating the DOM in the callback. We should avoid
reflowing in this case.